Scrollable window
Added a scrollable window to account for large number of found texture sets.
Tool also now automatically enables the "Triplanar" checkmark if "triplanar" is found anywhere in the filaname or path. 
This can be overridden by the user. 
Minor code fixes to prevent errors when user cancels out of the UI.